NTT: split NTT into separate file
